Split "task" related items into `silkenweb-task`
