+++ packages/task/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "silkenweb-task"
+version = { workspace = true }
+authors = ["Simon Bourne <simonbourne@gmail.com>"]
+edition = "2021"
+description = "Manage Silkenweb tasks"
+readme = "README.md"
+homepage = "https://github.com/silkenweb/silkenweb"
+repository = "https://github.com/silkenweb/silkenweb"
+license = "MIT OR Apache-2.0"
+
+categories = ["gui", "web-programming"]
+keywords = ["reactive", "web", "browser", "ssr"]
+
+[dependencies]
+silkenweb-macros = { workspace = true }
+crossbeam = { workspace = true }
+futures-signals = { workspace = true }
+clonelet = { workspace = true }
+futures = { workspace = true }
+
+[target.'cfg(all(target_arch = "wasm32", target_os = "unknown"))'.dependencies]
+js-sys = { workspace = true }
+wasm-bindgen = { workspace = true }
+wasm-bindgen-futures = { workspace = true }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+tokio = { workspace = true, features = ["rt", "macros", "time"] }
