+++ packages/task/README.md
@@ -1,0 +1,3 @@
+# Silkenweb Task
+
+Manage [Silkenweb](http://github.com/silkenweb/silkenweb) tasks.
